refactor: split http-logger and http-dispatcher into separate modules
Keeps createLoggingInterceptor in http-logger (logging only) and
introduces http-dispatcher solely for installHttpDispatcher, which
wires EnvHttpProxyAgent with the logging interceptor. Each module
now has a single responsibility.
