+++ packages/backlog-utils/src/index.ts
@@ -46,4 +46,4 @@ export {
   getRepoRelativePath,
   parseBacklogRemoteUrl,
 } from "./git-context";
-export { installHttpLogger } from "./http-logger";
+export { installHttpDispatcher } from "./http-dispatcher";
